исправлена архитектура проекта
