+++ GigaChat/forms/App/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:GigaChat"
-             StartupUri="MainWindow.xaml">
+             StartupUri="D:\GIT\windows-client\GigaChat\forms\LoaderWindow\LoaderWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
