+++ GigaChat/forms/App/App.xaml.cs
@@ -4,11 +4,7 @@ using System.Windows;
 
 namespace GigaChat
 {
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
     public partial class App : Application
     {
     }
-
 }
