+++ GigaChat/forms/App/AssemblyInfo.cs
