+++ GigaChat/forms/App/LocalData.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GigaChat
+{
+    public class LocalData
+    {
+        public const string HTTP_URL = "http://10.242.223.170:8084/user/@me/channels?sort=asc&order=id&meta=true";
+        public const string CUSTOM_CHANNEL = "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQOJo5kR1W2vqrBEKsd2JozDPdlrXf2Luz2Dg&usqp=CAU";
+        public const string DEFAULT_ICON_PATH = @"D:\GIT\windows-client\resourses\zeroImage.png";
+        public const long CLIENT_ID = 5;
+        public const string CLIENT_SECRET = "";
+        public const string CLIENT_KEY = "Et9pMkeTo9AYVCeDmzEiLmaHxS5kxtvkqQAoXiGNnfR7nzX9";
+    }
+}
