+++ GigaChat/forms/LoaderWindow/LoaderWindow.xaml
