+++ GigaChat/forms/MainWindow/BaseLogo.jpg
